requirements.txt
et-xmlfile==1.1.0
Faker==26.0.0
greenlet==3.0.3
mysql-connector-python==9.0.0
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
